+++ test/package/rules/babel.test.js
@@ -141,6 +141,14 @@ if (!babelConfig) {
       await compiler.run()
       expect(tracked[included]).toBeTruthy()
     })
+
+    test("explicitly included .cjs files should be transpiled", async () => {
+      const included = `${pathToNodeModulesIncluded}/a.cjs`
+      const [tracked, loader] = createTrackLoader()
+      const compiler = createTestCompiler(createWebpackConfig(included, loader))
+      await compiler.run()
+      expect(tracked[included]).toBeTruthy()
+    })
   })
 } // end of else block for babelConfig check